backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
PyPDF2==3.0.1
pdfplumber==0.11.7
openai>=1.0.0
python-dotenv==1.0.0
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
mangum==0.17.0

